pyproject.toml
[project]
name = "kazpunct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "datasets>=4.8.3",
    "fastapi>=0.135.1",
    "pytorch-crf>=0.7.2",
    "scipy>=1.17.1",
    "seqeval>=1.2.2",
    "transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
